+++ Driver/LdapDriverException.php
@@ -2,7 +2,7 @@
 
 namespace FR3D\LdapBundle\Driver;
 
-class LdapConnectionException extends \Exception
+class LdapDriverException extends \Exception
 {
 
     public function __construct($message)